Update Filter for team to appear in ranking from matches played to matches won
In the past (pre-2025), tournaments could invite teams they wanted, even if they where not yet ranked on the VRS. This allowed new teams with semi popular players (not good enough for Wildcard invites, but well known in the Tier 2/3 scene) to enter the ranking by playing 10 games or more. For those teams open qualifiers are now the only possiblity to enter the ranking. Winning one open qualifier will most likely not be enough to reach 10 games even with the following closed qualifier. Even winning two could not be enough. And the number of open qualifiers are limited. 

Thus in 2025, the filter for when a team needs to be in the ranking should be reduced from 10 to a lower number (e.g. 5). Additionally it would make sense to also base this filter on merit and not just on played games. My suggestion is thus, to change the filter to number of games won instead. My suggested value would be 3 games won.