more recent alternative match sample
Using the liquipedia API, manual cross reference and manual tweaks an alternative match sample has been created. This enables tweaking of the VRS formula using data representative of the current climate. This also includes a tweak to the data loader to correctly handle limited scope events which have greater coverage on liquipedia. This isnt a perfect recreation but is pretty close to the current sample