Updating VRS Model
Event stages that can be connected to later event stages (e.g., qualifiers or group stages) now add to their prize pool a portion of the final stage's prize pool.

Added a minimum participation requirement (5 matches of the past 10) for a player to appear on the active roster for a team on the VRS. Note that as a result, some rosters may not have five players.